LayoutTests/js/constant-encoding-expected.txt

Test that we correctly encode patterned immediate values

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 0 >= 0x01000100 is false
PASS 0 >= 0x01010000 is false
PASS 0 >= 0x00000101 is false
PASS 0 >= 0x00010001 is false
PASS 0 >= 0x01010101 is false
PASS successfullyParsed is true

TEST COMPLETE
